requirements.txt
et-xmlfile==1.1.0
lxml==4.9.1
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
six==1.16.0
